Update integration tests to account for host address containing port.
Since host.address now includes the port number some of the 'long'
integration tests needed to be updated to account for this as they used
the last character of the host address to validate expected hosts.

Introduced helper.lastOctetOf for retrieving the last octet of an
address or host.address.