Reduce number of configurations to build in jenkins on commit (#199)
On commit for all branches create a 'per_commit' job that runs the
following configurations:

  * nodejs 0.10 with latest C*
  * nodejs 6 with C* 1.2 and latest

On primary branches, create a 'nightly' job that runs all
configurations nightly if there were any changes.

On non-primary branches create a 'adhoc' job that runs all
configurations, but is not scheduled.  This job can
be used for triggering manual builds.